(int)GetFloatをGetIntに置き換え

+++ SuperNewRoles/Mode/SuperHostRoles/RPCSetTask.cs
@@ -26,7 +26,7 @@ namespace SuperNewRoles.Mode.SuperHostRoles
                     return false;
                 } else if (player.IsRole(RoleId.Workperson))
                 {
-                    var tasks = ModHelpers.GenerateTasks((int)CustomOptions.WorkpersonCommonTask.GetFloat(), (int)CustomOptions.WorkpersonShortTask.GetFloat(), (int)CustomOptions.WorkpersonLongTask.GetFloat()).ToArray();
+                    var tasks = ModHelpers.GenerateTasks(CustomOptions.WorkpersonCommonTask.GetInt(), CustomOptions.WorkpersonShortTask.GetInt(), CustomOptions.WorkpersonLongTask.GetInt()).ToArray();
                     MessageWriter messageWriter = AmongUsClient.Instance.StartRpc(player.NetId, (byte)29);
                     messageWriter.Write(playerId);
                     messageWriter.WriteBytesAndSize(tasks);
